refactor!: Remove AbstractCommentProcessor implementations and update CommentProcessorFactory
Removal of all AbstractCommentProcessor implementations has taken place to further optimize and streamline the codebase. Possible effects on CommentProcessorFactory have also been accounted for by incorporating the necessary updates. This effort focuses on improving readability, maintainability, and overall code organization.
