--- engine/src/main/java/org/wickedsource/docxstamper/processor/repeat/IRepeatProcessor.java
@@ -1,23 +1,0 @@
-package org.wickedsource.docxstamper.processor.repeat;
-
-import java.util.List;
-
-/**
- * Interface for processors that can repeat a table row.
- *
- * @author Joseph Verron
- * @author Tom Hombergs
- * @version ${version}
- * @since 1.0.0
- */
-public interface IRepeatProcessor {
-
-    /**
-     * May be called to mark a table row to be copied once for each element in the passed-in list.
-     * Within each copy of the row, all expressions are evaluated against one of the objects in the list.
-     *
-     * @param objects the objects which serve as context root for expressions found in the template table row.
-     */
-    void repeatTableRow(List<Object> objects);
-
-}
